abstract class for repetitive tasks
