+++ classes/suggested-tasks/local-tasks/providers/class-core-update.php
@@ -11,7 +11,7 @@ use Progress_Planner\Suggested_Tasks\Local_Tasks\Local_Task_Factory;
 /**
  * Add tasks for Core updates.
  */
-class Core_Update extends Local_Tasks_Abstract {
+class Core_Update extends Local_Repetitive_Tasks_Abstract {
 
 	/**
 	 * The capability required to perform the task.
@@ -41,52 +41,19 @@ class Core_Update extends Local_Tasks_Abstract {
 	 *
 	 * @return bool|string
 	 */
-	public function evaluate_task( $task_id ) {
-
-		// Early bail if the user does not have the capability to update the core, since \wp_get_update_data()['counts']['total'] will return 0.
-		if ( ! $this->capability_required() ) {
-			return false;
-		}
-
-		// Without this \wp_get_update_data() might not return correct data for the core updates (depending on the timing).
-		if ( ! function_exists( 'get_core_updates' ) ) {
-			require_once ABSPATH . 'wp-admin/includes/update.php'; // @phpstan-ignore requireOnce.fileNotFound
-		}
-
-		$task_object = ( new Local_Task_Factory( $task_id ) )->get_task();
-		$task_data   = $task_object->get_data();
-
-		if ( $task_data['type'] === static::ID && \gmdate( 'YW' ) === $task_data['year_week'] && 0 === \wp_get_update_data()['counts']['total'] ) {
-			return $task_id;
-		}
-		return false;
-	}
 
 	/**
-	 * Get an array of tasks to inject.
+	 * Check if the task condition is met.
 	 *
-	 * @return array
+	 * @return bool
 	 */
-	public function get_tasks_to_inject() {
-
-		// Early bail if the user does not have the capability to update the core or if the task is snoozed.
-		if ( true === $this->is_task_type_snoozed() || ! $this->capability_required() ) {
-			return [];
-		}
-
+	public function check_task_condition() {
 		// Without this \wp_get_update_data() might not return correct data for the core updates (depending on the timing).
 		if ( ! function_exists( 'get_core_updates' ) ) {
 			require_once ABSPATH . 'wp-admin/includes/update.php'; // @phpstan-ignore requireOnce.fileNotFound
 		}
 
-		// If all updates are performed, do not add the task.
-		if ( 0 === \wp_get_update_data()['counts']['total'] ) {
-			return [];
-		}
-
-		return [
-			$this->get_task_details( static::ID . '-' . \gmdate( 'YW' ) ),
-		];
+		return 0 === \wp_get_update_data()['counts']['total'] ? true : false;
 	}
 
 	/**
@@ -98,12 +65,16 @@ class Core_Update extends Local_Tasks_Abstract {
 	 */
 	public function get_task_details( $task_id ) {
 
+		if ( ! $task_id ) {
+			$task_id = $this->get_provider_id() . '-' . \gmdate( 'YW' );
+		}
+
 		return [
 			'task_id'     => $task_id,
 			'title'       => \esc_html__( 'Perform all updates', 'progress-planner' ),
 			'parent'      => 0,
 			'priority'    => 'high',
-			'type'        => static::TYPE,
+			'type'        => $this->get_provider_type(),
 			'points'      => 1,
 			'url'         => $this->capability_required() ? \esc_url( \admin_url( 'update-core.php' ) ) : '',
 			'description' => '<p>' . \esc_html__( 'Perform all updates to ensure your website is secure and up-to-date.', 'progress-planner' ) . '</p>',
